diploma/diploma/backend/requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
timm
scikit-learn
opencv-python
tqdm
pandas
datasets
